pyproject.toml
[tool.poetry]
name = "transactions_analyzer"
version = "0.1.0"
description = "Aplicação para análise de transações bancárias"
authors = ["Lucas Praciano <luskcct@gmail.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.9"
atomicwrites = "1.4.0"
attrs = "21.4.0"
blinker = "1.4"
click = "8.1.2"
colorama = "0.4.4"
Flask = "2.0.2"
Flask-Mail = "0.9.1"
greenlet = "1.1.2"
gunicorn = "20.1.0"
iniconfig = "1.1.1"
itsdangerous = "2.1.2"
Jinja2 = "3.0.3"
jsonschema = "4.5.1"
MarkupSafe = "2.0.1"
marshmallow = "3.15.0"
packaging = "21.3"
pluggy = "1.0.0"
psycopg2 = "2.9.2"
py = "1.11.0"
PyJWT = "2.3.0"
pyparsing = "3.0.8"
pyrsistent = "0.18.1"
python-dotenv = "0.20.0"
six = "1.16.0"
SQLAlchemy = "1.4.28"
SQLAlchemy-Utils = "0.38.2"
tomli = "2.0.1"
Werkzeug = "2.0.2"

[tool.poetry.dev-dependencies]
blue = "^0.9.0"
pytest = "^7.1.2"
pip-audit = "^2.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
